% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_to_dataframe.R
\name{as.data.frame.redcap_csv}
\alias{as.data.frame.redcap_csv}
\title{FUNCTION_TITLE}
\usage{
\method{as.data.frame}{redcap_csv}(x, ...)
}
\arguments{
\item{x}{PARAM_DESCRIPTION}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\seealso{
\code{\link[httr2]{resp_body_raw}}
}
