% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_record.R
\name{check_record}
\alias{check_record}
\title{FUNCTION_TITLE}
\usage{
check_record(study, data)
}
\arguments{
\item{study}{PARAM_DESCRIPTION}

\item{data}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\seealso{
\code{\link[dplyr]{rename}}, \code{\link[dplyr]{distinct}}, \code{\link[dplyr]{across}}, \code{\link[dplyr]{reexports}}, \code{\link[dplyr]{mutate-joins}}, \code{\link[dplyr]{select}}
}
