% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_metadata.R
\name{export_metadata}
\alias{export_metadata}
\title{export_metadata}
\usage{
export_metadata(study, format = c("csv", "json"), forms = NULL, fields = NULL)
}
\arguments{
\item{study}{A study acronym or name to select the token from a user's R environment (stored as STUDY_TOKEN)}

\item{format}{The file type to retrieve from REDCap (choices include csv or json), Default: csv}

\item{forms}{An array of form names specifying specific data collection instruments for which you wish to pull
metadata (by default, all metadata is pulled), Default: NULL}

\item{fields}{An array of field names specifying specific fields you wish to pull (by default, all metadata is pulled), Default: NULL}
}
\value{
A data frame containing the codebook from the REDCap project
}
\description{
Exports the metadata or codebook from the REDCap API
}
\details{
Exports the metadata or codebook from the REDCap API
}
\seealso{
\code{\link[httr2]{request}}, \code{\link[httr2]{req_body}}, \code{\link[httr2]{req_perform}}
}
