% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_record.R
\name{export_record}
\alias{export_record}
\title{FUNCTION_TITLE}
\usage{
export_record(
  study,
  records = NULL,
  forms = NULL,
  fields = NULL,
  events = NULL,
  format = c("csv", "json"),
  rawOrLabel = c("raw", "label"),
  rawOrLabelHeaders = c("raw", "label"),
  exportCheckboxLabel = "false",
  type = c("flat", "eav"),
  ...
)
}
\arguments{
\item{study}{A study acronym or name to select the token from a user's R environment (stored as STUDY_TOKEN)}

\item{records}{PARAM_DESCRIPTION, Default: NULL}

\item{forms}{PARAM_DESCRIPTION, Default: NULL}

\item{fields}{PARAM_DESCRIPTION, Default: NULL}

\item{events}{PARAM_DESCRIPTION, Default: NULL}

\item{format}{PARAM_DESCRIPTION, Default: c("csv", "json")}

\item{rawOrLabel}{PARAM_DESCRIPTION, Default: c("raw", "label")}

\item{rawOrLabelHeaders}{PARAM_DESCRIPTION, Default: c("raw", "label")}

\item{exportCheckboxLabel}{PARAM_DESCRIPTION, Default: 'false'}

\item{type}{PARAM_DESCRIPTION, Default: c("flat", "eav")}

\item{...}{PARAM_DESCRIPTION}
}
\value{
OUTPUT_DESCRIPTION
}
\description{
FUNCTION_DESCRIPTION
}
\details{
DETAILS
}
\examples{
\dontrun{
if(interactive()){
 #EXAMPLE1
 }
}
}
\seealso{
\code{\link[httr2]{request}}, \code{\link[httr2]{req_body}}, \code{\link[httr2]{req_perform}}
}
