% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_reports.R
\name{export_reports}
\alias{export_reports}
\title{export_reports}
\usage{
export_reports(
  study,
  report_id,
  format = c("csv", "json"),
  rawOrLabel = c("raw", "label"),
  rawOrLabelHeaders = c("raw", "label"),
  exportCheckboxLabel = "false"
)
}
\arguments{
\item{study}{A study acronym or name to select the token from a user's R environment (stored as STUDY_TOKEN)}

\item{report_id}{A numeric value representing the REDCap report to be downloaded to the user's R environment}

\item{format}{The file type to retrieve from REDCap (choices include csv or json), Default: csv}

\item{rawOrLabel}{Export the raw coded values or labels for the options of multiple choice fields, Default: raw}

\item{rawOrLabelHeaders}{Export the variable or field names (raw) or the field labels (label). Labels can only
be exported if format is csv, Default: raw}

\item{exportCheckboxLabel}{Specifies the format of checkbox when exporting the data as labels (i.e., when
rawOrLabel = "label"). Setting exportCheckboxLabel to "true" will add the checkbox value instead of using
'Checked' or 'Unchecked'.}
}
\value{
A data frame containing the extracted report
}
\description{
Retrieves a report from the REDCap API
}
\details{
Retrieves a report from the REDCap API
}
\seealso{
\code{\link[RCurl]{postForm}}
\code{\link[jsonlite]{toJSON, fromJSON}}
}
