% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_prompt_schedule.R
\name{generate_prompt_schedule}
\alias{generate_prompt_schedule}
\title{generate_prompt_schedule}
\usage{
generate_prompt_schedule(
  id,
  repeat_instrument = "ema",
  enrollment_date = Sys.Date(),
  n_days = 7,
  break_length = c(),
  prompts_per_day = 4,
  time_blocks = c(8, 11, 14, 18, 21),
  add_prompt_time = TRUE,
  random_surveys = NULL,
  random_survey_probability = 0.5,
  event_names = NULL,
  availability = NULL,
  ...
)
}
\arguments{
\item{id}{The record_id that uniquely identifies each participant}

\item{repeat_instrument}{The name of the repeating instrument containing the EMA questions, Default: 'ema'}

\item{enrollment_date}{The participant's enrollment date to generate the dates for the schedule, Default: Sys.Date()}

\item{n_days}{The number of days included in the prompting schedule. The n_days argument can be a single number
or an integer vector to represent multiple sampling periods, Default: 7}

\item{break_length}{The length of time in between sampling periods if n_days is an integer vector of
multiple sampling periods, The break_length must be one less than the length of n_days Default: c()}

\item{prompts_per_day}{The number of prompts per day. The length of the prompts_per_day argument should
match that of n_days, Default: 4}

\item{time_blocks}{The 24-hour time blocks for controlling when the prompts are sent. Time blocking is used
to ensure the prompts are delivered sequentially throughout the day, but a random time will be generated
within each of the time blocks. The time_blocks argument should be a length one more than n_days, Default: c(8, 11, 14, 18, 22)}

\item{add_prompt_time}{Add a random prompt time variable to the data set within the time_blocks, Default: TRUE}

\item{random_surveys}{A selection of surveys to deliver randomly. This generates a controller variable that delivers different
combinations of the surveys ranging from none to all that are sampled at random_survey_probability, Default: NULL}

\item{random_survey_probability}{A random sampling probability used to generate the controller to designate a random
delivery schedule for the designated surveys. Single random_survey_probability arguments are repeated to match the
length of random_surveys, Default: 0.5}

\item{event_names}{The name of the events to inclue in the data set that coincide with each sampling period. If provided, event_names
should be the same length as the n_days argument and map onto redcap_event_name., Default: NULL}

\item{...}{Optional arguments for a secondary controller in the form of a named list matching one of the random_surveys. Within the list,
users can use the c() operator to control the sampling strategy in that one value within c() will be picked each time random_surveys
are sampled.}
}
\value{
A data frame that can be passed to import_record to set up the prompting schedule.
}
\description{
Generates a prompt schedule for an EMA study that can be hosted on REDCap
}
\details{
Generates a prompt schedule for an EMA study that can be hosted on REDCap
}
\seealso{
\code{\link[purrr]{flatten}}, \code{\link[purrr]{map2}}, \code{\link[purrr]{map}}
}
