% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_record.R
\name{import_record}
\alias{import_record}
\title{import_record}
\usage{
import_record(
  study,
  format = "json",
  type = c("flat", "eav"),
  data = NULL,
  ...
)
}
\arguments{
\item{study}{The study name or abbreviation that matches the API token stored in a user's R environment.
The token should be stored as the study name or abbreviation followed by _TOKEN.}

\item{format}{The type of format to be uploaded to REDCap. Choices can be csv or json; however,
only json is implemented at this time., Default: "json"}

\item{type}{Output is one record per row (flat) or one data point per row (eav), Default: flat}

\item{data}{A vector or data frame containing data to upload to REDCap, Default: NULL}

\item{...}{Individual variables from the REDCap database to be uploaded in the format of variable = "value".
A user can pass as many variables as needed to ..., but they are only used if data is NULL}
}
\value{
The status of the request
}
\description{
Imports a record to REDCap through the API
}
\details{
Imports a record to REDCap through the API
}
\seealso{
\code{\link[jsonlite]{toJSON, fromJSON}}, \code{\link[jsonlite]{prettify, minify}}
\code{\link[httr2]{request}}, \code{\link[httr2]{req_body}}, \code{\link[httr2]{req_perform}}
}
